modify default to vector<double>

+++ scripts/timer.cpp
@@ -2,7 +2,7 @@
 #include<vector>
 using namespace std;
 
-template<typename T0, typename T1=vector<int> >
+template<typename T0, typename T1=vector<double> >
 void timer(T0 func, T1 arg) {
 
   *func << &arg;
